+++ src/pythonequipmentdrivers/daq/__init__.py
@@ -1,4 +1,4 @@
-from .agilent_34972a import Agilent_34972A
+from ._agilent_34972a import Agilent_34972A
 
 __all__ = [
     "Agilent_34972A"
