+++ src/pythonequipmentdrivers/daq/_agilent_34972a.py
