+++ src/pythonequipmentdrivers/functiongenerator/__init__.py
@@ -1,5 +1,5 @@
-from .agilent_33250a import Agilent_33250A
-from .keysight_33500b import Keysight_33500B
+from ._agilent_33250a import Agilent_33250A
+from ._keysight_33500b import Keysight_33500B
 
 __all__ = [
     "Keysight_33500B",
