+++ src/pythonequipmentdrivers/functiongenerator/_agilent_33250a.py
