+++ src/pythonequipmentdrivers/functiongenerator/_keysight_33500b.py
