+++ src/pythonequipmentdrivers/multimeter/__init__.py
@@ -1,8 +1,8 @@
-from .fluke_45 import Fluke_45
-from .fluke_8845a import Fluke_8845A
-from .fluke_dmm import Fluke_DMM
-from .hp_34401a import HP_34401A
-from .keysight_34461a import Keysight_34461A
+from ._fluke_45 import Fluke_45
+from ._fluke_8845a import Fluke_8845A
+from ._fluke_dmm import Fluke_DMM
+from ._hp_34401a import HP_34401A
+from ._keysight_34461a import Keysight_34461A
 
 __all__ = [
     "Fluke_45",
