+++ src/pythonequipmentdrivers/multimeter/_keysight_34461a.py
@@ -1,4 +1,4 @@
-from .hp_34401a import HP_34401A
+from ._hp_34401a import HP_34401A
 
 # TODO: Add in additional measurement functionallity not written in the
 # original HP_34401A Class. Look into adding functionallity for adjusting the
