+++ src/pythonequipmentdrivers/networkanalyzer/__init__.py
@@ -1,4 +1,4 @@
-from .bode100 import Bode100
+from ._bode100 import Bode100
 
 __all__ = [
     "Bode100",
