+++ src/pythonequipmentdrivers/oscilloscope/__init__.py
@@ -1,6 +1,6 @@
-from .lecroy_wr8xxx import Lecroy_WR8xxx
-from .tektronix_dpo4xxx import Tektronix_DPO4xxx
-from .tektronix_mso5xxx import Tektronix_MSO5xxx
+from ._lecroy_wr8xxx import Lecroy_WR8xxx
+from ._tektronix_dpo4xxx import Tektronix_DPO4xxx
+from ._tektronix_mso5xxx import Tektronix_MSO5xxx
 
 __all__ = [
     "Lecroy_WR8xxx",
