+++ src/pythonequipmentdrivers/oscilloscope/_lecroy_wr8xxx.py
