+++ src/pythonequipmentdrivers/oscilloscope/_tektronix_dpo4xxx.py
