+++ src/pythonequipmentdrivers/oscilloscope/_tektronix_mso5xxx.py
