+++ src/pythonequipmentdrivers/source/__init__.py
@@ -1,16 +1,16 @@
 from typing import Protocol
 
-from .agilent_6030a import Agilent_6030A
-from .bkprecision_9132b import BKPrecision_9132B
-from .californiainstruments_csw5550 import CaliforniaInstruments_CSW5550
-from .chroma_62000p import Chroma_62000P
-from .elgar_1750a import Elgar_1750A
-from .elgar_5250a import Elgar_5250A
-from .hp_6632a import HP_6632A
-from .intepro_psi9000 import Intepro_PSI9000
-from .keithley_2231a import Keithley_2231A
-from .ppsc_3150afx import PPSC_3150AFX
-from .sorensen_sga import Sorensen_SGA
+from ._agilent_6030a import Agilent_6030A
+from ._bkprecision_9132b import BKPrecision_9132B
+from ._californiainstruments_csw5550 import CaliforniaInstruments_CSW5550
+from ._chroma_62000p import Chroma_62000P
+from ._elgar_1750a import Elgar_1750A
+from ._elgar_5250a import Elgar_5250A
+from ._hp_6632a import HP_6632A
+from ._intepro_psi9000 import Intepro_PSI9000
+from ._keithley_2231a import Keithley_2231A
+from ._ppsc_3150afx import PPSC_3150AFX
+from ._sorensen_sga import Sorensen_SGA
 
 
 class VoltageSource(Protocol):
