+++ src/pythonequipmentdrivers/source/_bkprecision_9132b.py
@@ -1,4 +1,4 @@
-from .keithley_2231a import Keithley_2231A
+from ._keithley_2231a import Keithley_2231A
 
 
 # acts as an alias of Keithley_2231A
