+++ src/pythonequipmentdrivers/temperaturecontroller/__init__.py
@@ -1,4 +1,4 @@
-from .koolance_exc900 import Koolance_EXC900
+from ._koolance_exc900 import Koolance_EXC900
 
 __all__ = [
     "Koolance_EXC900",
